+++ MHD/C_src/KEMO_GL/m_kemoview_fline.h
@@ -38,7 +38,7 @@ void set_fline_parameters(int selected, int input, struct kemoview_fline *kemo_f
 int get_fline_parameters(struct kemoview_fline *kemo_fline, int selected);
 
 void set_fline_field_param(int selected, int input, struct kemoview_fline *kemo_fline);
-int get_fline_field_param(int selected, struct kemoview_fline *kemo_fline);
+long get_fline_field_param(int selected, struct kemoview_fline *kemo_fline);
 
 void set_fline_color_param(int selected, int input, struct kemoview_fline *kemo_fline);
 int get_fline_color_param(int selected, struct kemoview_fline *kemo_fline);
