+++ MHD/C_src/KEMO_GL/set_each_fline_parameters.h
@@ -34,21 +34,21 @@ void set_fline_switch(struct fline_menu_val *fline_m, int iflag);
 void set_fline_color_type(struct fline_menu_val *fline_m, int iflag);
 
 int get_fline_switch(struct fline_menu_val *fline_m);
-int get_fline_color_num_field(struct psf_data *fline_d);
-int get_fline_color_ncomptot(struct psf_data *fline_d);
+long get_fline_color_num_field(struct psf_data *fline_d);
+long get_fline_color_ncomptot(struct psf_data *fline_d);
 int fline_color_num_comps(struct psf_data *fline_d, int i);
 int get_fline_color_istack(struct psf_data *fline_d, int i);
 void get_fline_color_data_name(struct psf_data *fline_d, 
 			struct kv_string *colorname, int i);
-int get_fline_color_field(struct fline_menu_val *fline_m);
-int get_fline_color_component(struct fline_menu_val *fline_m);
-int get_fline_color_data_adress(struct fline_menu_val *fline_m);
+long get_fline_color_field(struct fline_menu_val *fline_m);
+long get_fline_color_component(struct fline_menu_val *fline_m);
+long get_fline_color_data_adress(struct fline_menu_val *fline_m);
 int get_fline_colormode(struct fline_menu_val *fline_m);
 
 
-void set_fline_type(struct fline_menu_val *fline_m, int iflag);
-int get_fline_type(struct fline_menu_val *fline_m);
-int toggle_fline_type(struct fline_menu_val *fline_m);
+void set_fline_type(struct fline_menu_val *fline_m, long iflag);
+long get_fline_type(struct fline_menu_val *fline_m);
+long toggle_fline_type(struct fline_menu_val *fline_m);
 
 void set_fline_thickness(double value, struct fline_menu_val *fline_m);
 double get_fline_thickness(struct fline_menu_val *fline_m);
